+++ bobtemplates/plone/behavior.py
@@ -5,8 +5,8 @@ from bobtemplates.plone.base import git_commit
 from bobtemplates.plone.base import update_file
 from lxml import etree
 
+import case_conversion as cc
 import os
-import stringcase
 
 
 def _update_package_configure_zcml(configurator):
@@ -87,9 +87,9 @@ def prepare_renderer(configurator):
     configurator = base_prepare_renderer(configurator)
     configurator.variables['template_id'] = 'behavior'
     behavior_name = configurator.variables['behavior_name'].strip('_')
-    configurator.variables['behavior_name_klass'] = stringcase.pascalcase(
+    configurator.variables['behavior_name_klass'] = cc.pascalcase(
         behavior_name)
-    configurator.variables['behavior_name_normalized'] = stringcase.snakecase(  # NOQA: E501
+    configurator.variables['behavior_name_normalized'] = cc.snakecase(  # NOQA: E501
         behavior_name)
     configurator.target_directory = configurator.variables['package_folder']